+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * Upgrade `rubocop` dependency from `0.15.0` to `0.16.0`
+* Fix broken `--show-linters` flag
 
 ## 0.3.0